Show logs in terminal (#101)
The logs of a Pod can now be accessed via the actions. Therefore the
logs tab was removed from the Pod details. When a user selects the logs
action a modal with various options will be shown. When the user then
clicks the "Show Logs" button the logs will be loaded and a new terminal
with the retrieved log lines will be added.