Add option to get user information from a request
This commit adds a new authentication middleware for the kobs API. This
middleware is used to get user information (e.g. the email address of an
user) from a HTTP header. The header can be set via the
"--api.auth-header" flag. If this header is present kobs will log all
requests including a user field.

We also added an example how this header can be passed to kobs, when
kobs is protected using the OAuth2 Proxy.

In the future we will use this feature to extend some plugins like the
Opsgenie plugin, where we can use this user for additional actions like
acknowledge or snooze alerts.