Add actions for Opsgenie plugin
With the new authentication logic, which was introduced in #103 it is
now possible to acknowledge, snooze and close Opsgenie alerts via kobs.
These action must be enabled via the configuration file for kobs. When
they are enabled they will be added as additional actions in the React
UI. If kobs is used with a setup like it was introduced in #103 we will
use the correct user for these action. If no user is set we use the
default user named "kobs.io".