Add Prometheus metrics for API requests
We are adding a new middleware, which is used to generate Prometheus
metrics for all API requests. The middleware generates two metrics for
each status, method and path: kobs_chi_requests_total and
kobs_chi_request_duration_milliseconds. These metrics can be used to
monitor the performance of all kobs API requests.

We also add an option to the kobs Helm chart to create a Service Monitor
for the Prometheus Operator, which can be used to collect these metrics.