Change Prometheus sparkline chart to allow labels
It is now possible to use the labels of a Prometheus metric as value in
the sparkline chart. This means when the user provides a label in a
sparkline chart the label will be shown instead of the current value of
the returned metric.

We also improved the calculation of the average value of a metric by
excluding all the null values.