Add new option for Prometheus chart legend
This adds a new option "table-large" for the Prometheus chart legend,
which can be used to have a larger legend table, when there are a lot of
values to display in the legend.

We also changed the formatting of the values in the legend and in
sparkline charts, so that they are rounded to a maximum of 4 decimals.