Improve tooltip position in all nivo charts (#108)
Until now it was possible that the tooltip in the nivo charts wasn't
completly visible. That should be fixed now. For that we are checking if
the point for which the tooltip is shown is in the first half or second
half of the chart. If it is in the first half the tooltip will be
displayed on the right of the cursor and if it is in the second half the
tooltip will be displayed on the right of the cursor.

We also apply the roundNumber function, which was introduced in #107, in
the table component of the Prometheus plugin.

The "/api/health" endpoint is now excluded from all middlewares. This
was changed, so that we do not spam metrics and logs with the values for
this enpoint, because it makes finding the right logs in a Kubernetes
environment with regular health check difficult.

Last but not least we updated all the NPM packages, which we are using
in the plugins.