Fix tooltip position in Prometheus charts (#109)
The tooltip position wasN't correct in charts where a metric contained
gabs. This is now fixed by setting the min and max value for the x axis
to the selected time range and by using this time range to determin if a
point is in the first or seconde half of the chart.