Fix dashboard tabs showing wrong variables
Until now it was possible that a dashboard shows the wrong variables /
the variables from another dashboard. This could happen when the user
selected a dashboard with a larger index before another dashboard and
then switched to a dashboard with a lower index.

To fix this issue we do not use the mountOnEnter property from the Tabs
component anymore. Instead we render all dashboards when the Dashboards
component is rendered. To not load all the data for all dashboards we
are passing the activeKey and eventKey from the Tabs component to the
Dashboard component and only load the variables, when the activeKey is
equal the eventKey.