Fix usage of memo and resource table for CRDs
This commit fixes the usage of "memo" in the Dashboard component, where
we forgot some important properties.

This also fixes a problem with CRDs in the resources table. When the
JSONPath for a shown column returned undefined and the column was of
type date it showed NaN in the table. Now we are returning an empty
string when the returned value from the JSONPath is undefined.