Add dashboard title as query key for variables
This commit adds the dashboard title as query key for the API call to
load the variables for a dashboard. This is required, because when two
or more dashboards have the same or no variables, it will always show a
spinner for the seconde one, because all query keys are the same. This
is fixed by adding the dashboard title as a unique query key accross all
referenced dashboards.