Allow mapping values in Prometheus table panel (#112)
It is now possible to use the value mapping from the sparkline panel in
the table panel for the Prometheus plugin. To use this feature we added
a new option for the specified columns named "mapping". If a user
specifies such mappings the corresponding value will be shown in the
table instead of the returned value from the PromQL query.

We also using the "dashboard" as query key instead of the
"dashboard.title".

Last but not least it is now possible to filter the variable values in a
dashboard.