Improve and allow customization of axis scaling
It is now possible to customize the scaling of the y axis via the new
"yAxis" option for the Prometheus plugin. A user can set the minimum and
maximum value for the axis. A user can also specify two special string
values "auto" or "min" / "max" to automatically scale the axis or to use
the minimum / maximum value accross all loaded metrics.

We also improved the scale of the x axis by using the selected time
range or if there is a lower / higher value in the metrics we are using
this value instead, so that the lines / areas in a chart not rendered
outside of the specified range.