Fix span order and additional fields for Jaeger (#114)
The Jaeger plugin had two bugs, where spans were not shown in the
correct order, because we didn't use the start time of the root span,
but the start time of the first span in a trace and we didn't sorted the
spans by there start date before we build the our span tree. These two
bugs are fixed now.

We also fixed a bug, where the additional fields from the options modal
on the Jaeger page, where not used to get the list of traces, because we
didn't changed the additional fields data in the toolbar component.

We also improved the loading indication for several plugins by not
keeping the previous data. This feature is now only used for the
dashboard variables and the Prometheus plugin, where we show the spinner
in the panel header when new data is loaded.

Last but not least we improved the time handling in the Options
component of the core package, so that we keep the selected time range
when a user only changes an additional field.