Fix Prometheus charts, when metrics are null (#118)
When the API call to get the metrics for a Prometheus chart returned
null an error "null is not an object (evaluating 'e[Symbol.iterator]')"
was thrown. This is now fixed, by checking if the metrics object is null
and if this is the case we are returning an empty array instead of null
in the data. This was already handled for the sparkline charts.

We also showing all the dependencies of an application in the link list
in the application details.