Add Flux Plugin (#119)
The Flux plugin can be used to view and reconcile all Flux resources
like Git Repositories, Helm Repositories, Kustomizations and Helm
Releases. These resources can be accessed via the Flux page or they can
be embedded into a dashboard via the Flux panel.

To use the panel a user must set the "type" for the resources which
should be displayed. He also can set the optional values for the
cluster, namespace and selector. The selector will select the Flux
resources by the specified label "lable=value".

We also added a fix for the Kiali documentation, which indicated a wrong
status of the Kiali plugin. The Kiali plugin is production ready and all
planned features are implemented.