Fix reconciliation of Flux resources
The Flux plugin contains a feature to trigger a reconciliation of
Kustomizations and HelmReleases. This feature didn't worked as expected,
because a patch of the corresponding annotation didn't triggered the
reconciliation.

Now we are using a similar logic like it is used in the fluxcd/webui
repository to trigger the reconciliation of Kustomizations and
HelmReleases. For that we also had to add a new method for all clusters,
to return a client which can be used for custom resources.