Add ClickHouse Plugin
This is the first version for the ClickHouse plugin, which can currently
only be used to query logs from a ClickHouse instance ingested via the
kobsio/fluent-bit-clickhouse Fluent Bit plugin. For the next version we
plan to also add support for raw sql queries.

When the plugin is run in the "logs" mode, which can be set via the
"type" option it can be used to get ingested logs from a ClickHouse
instance via our custom query language. The user just provides a query
and a time range and we parse the given query to have a valid where
statement for the sql query. Then we are returning the ingested logs as
json documents.

We also adjusted the core package to pass the additonal options for a
plugin to the plugin page. This was required, because we have to decide
which view a user should see for the specified "type" in the ClickHouse
plugin configuration.