Allow multiple queries for Elasticsearch
It is now possible to specify multiple queries in the panel options for
the Elasticsearch plugin. For that a new field queries was added to the
options, which can be used to set a name, the Elasticsearch query and a
list of fields. If the user sets more then one query he can select the
query for which the results should be shown via a select box in the
frontend.

This is a breaking change for the Elasticsearch plugin. This means that
the old fields query and fields are not working anymore with the
Elasticsearch plugin.

We also fixed a bug, which causes the Elasticsearch page to crash, when
no documents where returned for a query.