Show request details for ClickHouse logs (#126)
We are now showing the details for log requests in the ClickHouse
plugin. We added a card header to provide information like the number of
all documents for the query and the time it took to get the entries from
ClickHouse.

This commit also fixes a small typo in the query language parsing, where
closing brackets were not recognized. This is now fixed and queries with
brackets should work as expected.

Last but not least, we are showing a small spinner in the header of the
card for logs retrieved in the ClickHouse or Elasticsearch plugin. So
that a user see that his query is still executed after a change to the
time range or query.