Specify Dashboards within Teams and Applications
It is now possible to specify Dashboards within a Team or Application
via the new "inline" property. This can be used if the Dashboard should
only be used within the Team or Application and isn't required for other
resources.

The inline property takes a list of variables and rows. Placeholders are
not possible within these inline definitions. If a user provides the
cluster/namespace/name for a Dashboard and the inline property, we
always use the inline property.