Fix handling of traces in the Jaeger plugin (#129)
We had the problem, that in some cases the UI of the Jaeger plugin was
broken, when the trace wasn't as expected. To handle all these special
cases we decided to use some functions from the jaegertracing/jaeger-ui
repository, because the know what they do. Now all cases which can be
handled by the Jaeger UI should also work within kobs.

We also improved they way how we are rendering the list of spans for a
trace. in former versions of kobs, the UI crashed when a trace contained
to much spans. Now we are using react-virtuoso to render these large
lists of spans, which improves the performance quite a lot. We tested
this with a trace with 13500 spans. The current solution for rendering
the spans isn't perfect, because we had to set a fixed sice for the
container. In the future we want to use "useWindowScroll" property of
the package, but at the moment we are seeing a bug with react-virtuoso
as it is also described in the react-virtualized project (#1671).