Allow multiple queries for Jaeger
Some days ago we change the Elasticsearch plugin to support multiple
queries in the panel. In the same way we are chaning the options for the
Jaeger plugin now. So that a user can specify multiple queries within a
Jaeger panel. In the UI the user can then select the query he wants to
see via the select box.

This is a breaking change for the Jaeger panel, this means that old
specification wont work anymore with the current version of kobs.