Add chart for distribution of logs in ClickHouse
This commit adds a new chart component, which shows the distribution of
logs in the selected time range for the ClickHouse plugin, when it is
used in the "logs" mode. With the new chart it is possible to see when
the most of the logs were written in the selected time range. The chart
always contains 30 buckets and is simular to the one, which is known
from Kibana.

The chart can also be used within a ClickHouse panel. For this a new
option "showChart" was added. This option is only used when the type for
the panel is set to "logs".

We also adjusted the CodeQL to remove some warnings and following the
latest recommendations.