Improve query performance for ClickHouse (#133)
Queries to get logs from ClickHouse over a large time range should now
be faster. For this we are only returning a maximum of round about 10000
documents from ClickHouse. For this we had to remove the stats enpoint
which returned the overall document count and the buckets data and added
it to the documents endpoint. With the help of the retrieved buckets we
are modifying the start time of a query, so that large queries should be
faster. More information about this new approach can be found in the
code as inline comment.

For example, queries to get all logs from the last 7 days, which took up
to 3 minutes, are taking 10 seconds now.

Note: We decided for the 10000 limit, because the default limit in
Kibana is 500 and so it should be large enough. In the future we can
also provide an option in the ClickHouse configuration or an additional
field in the Options component to increase this limit.