Allow custom order for returned logs in ClickHouse
It is now possible to set a custom order for the retunred logs from the
user defined query. For that two new options "order" and "orderBy" were
added which allows a user to specify if the logs should be sorted
ascending or descending and to set the fields by which the logs should
be ordered.

We also added two new operators. The "!~" operator can be used for "NOT
ILIKE" queries and the "_exists_" operator can be used to filter out
NULL values.

We also improved the UI, so that the details view for a log line isn't
opened by clicking on the row. Instead we added a action dropdown which
can be used to open the details view for a log line.