Change log view for ClickHouse and Elasticsearch
We are changing the log view of the ClickHouse and Elasticsearch plugin
to improve the experience for developers. The new view is very simular
to the one known from Kibana. So instead of showing the details for a
log line in the drawer, the rows can now be expanded, to see the table
or json representation of a log line.

We also added a link to the documentation for the ClickHouse and
Elasticsearch plugin. The documentation can be found at kobs.io. This
should help user to find the query syntax for the plugins faster.

We also fixed a small bug for the ClickHouse plugin. The buckets chart
can not be created if the user selected a time range smaller than 30
seconds, because we also try to generate 30 buckets. This case is now
checked before the sql query is run.