Add option to stream (follow) logs via WebSockets (#138)
A new option follow was added for the logs of a container. By selecting
the "follow" option in the logs modal, we will establish a WebSocket
connection. Via this connection we will then stream all incoming logs of
the selected container and show them in the terminal. This is like the
"-f" option for kubectl.