Fix details row for ClickHouse and Elasticsearch
The details row which can be shown for logs from ClickHouse and
Elasticsearch used the wrong col span, because we missed to add the time
column in case the user selected some fields.

We had a copy and paste error for the distribution chart in the
Elasticsearch plugin, were we used the wrong key for the bottom axis of
the chart.