Fix bug where Jaeger results were not refreshed
In #130 we changed the options for the Jaeger plugin to allow multiple
queries. For that we set the initial selected query to the first query
from the "queries" property. This caused a bug, were the results were
not refreshed after a user changed the service, operation, etc. on the
Jaeger page, because we didn't updated the selected query. The selected
query is now always updated when the queries property is changed and
when it only contains one query, which is always the case for the page
implementation of the Jaeger plugin.

Fixes #142.