Improve query performance for ClickHouse plugin
Instead of running the counts and buckets query before getting the logs
from ClickHouse we are now just running the buckets query. The count
query was removed, because we can also determine the count by building
the sum of the count in each bucket. This should improve the query
performance for some queries by up to 50%.

In #133 we introduced an setting, where we limit the amount of
documents, which are returned by the API. There we set a limit of 10000
documents. This setting can now be set by the user via the additional
options. If the user doesn't provide this option we set a default limit
of 1000.