Improve reliability of kobs
This commit improves the reliability of kobs, be not checking the
database connection in the ClickHouse instance after a new instance was
created. In this way kobs can also start, when the configured ClickHouse
instance is temporarly unavailable.

We also fixed a small bug, where we forgot to stop the goroutine which
writes a new line for long running queries, when an error occured.