Add SQL Plugin (#149)
The new SQL plugin can be used to show the data from a SQL database
within kobs. In the first version we are only supporting the
"clickhouse", "postgres" and "mysql" driver in the SQL plugin.

We also only supporting to display the retrieved data in a table, later
we can also add other visualizations, like bar, pie or line charts.

NOTE: The plugin just runs the given query, which means if a user
proviedes a "DELETE" or "UPDATE" statement it will be executed by the
plugin. For that you should handle the permissions of the user which is
provided to access the database on the database layer.