Add actions for log details
This commit adds actions to the log details in the ClickHouse plugin, so
that users can filter for values based on the value of a log line and
that they can toggle a specific field directly from the details.

This commit also improves how we check if a field exists when it is not
a default field.