Improve performance for large dashboards (#152)
We are now using Intersection Observer API for dashboards to improve the
performance of large dashboards with a lot of panels. The panels are now
only rendered when the become visible in the viewport. This improves the
perfromance for large dashboards, because we render less DOM elements at
the same time.

When an Application in the gallery view is selected we are now going
directly to the corresponding Application page and do not show the
Application in a panel. The panel view is still used for the
applications which are selected in the topology view.