Improve query performance for ClickHouse plugin (#157)
Instead of just changing the start time for the raw logs query, we are
now running the raw logs query only for buckets which are containing
documents, to improve the query performance for queries which are
returning a low number of documents.

We also improved the performance of the React UI by using the
Intersection Observer API to render the table with the returned log
documents.

Finally we have removed the property to specify the maximum number of
documents and always returning 1000 documents. For that we also removed
the pagination / load more documents logic.