Improve HTTP metrics and add ClickHouse metric (#158)
Improve the exported HTTP metrics, by using summary instead of
histograms for the request latency.

Add a new ClickHouse metric, to determine how often a field is used in
queries. With the help of this metric we can then add more fields to the
logs table to improve the query time for the fields which are used most
of the time.