Allow users to select a time range in logs chart (#159)
It is now possible to select a new time range in the logs chart of the
ClickHouse plugin. This allows users to zoom into a smaller time range
to view the returned logs.

For this feature we switched from nivo charts to the
patternfly/react-charts charting package. This package comes with a
built in function for "brushing" which makes it easier to provide this
feature.

NOTE: For the charts in the other plugins like Prometheus and Jaeger we
can adjust the style of the charts to follow the styling from the
patternfly/react-charts package. This could provide a better overall
user experience for all charts.