Allow users to sort logs via the documents table (#160)
We are now showing the selected sort order in the table with the
returned logs in the ClickHouse plugin. Besides that a user can now also
change the order of the returned documents in this table.

This commit also fixes a bug in the logs chart, where the chart wasn't
rendered, when it was used in a dashboard.