Add support for materialized columns (#161)
It is now possible to set a list of materialized columns in the
ClickHouse plugin configuration. This allows users to speed up queries
for the most frequently used fields. The columns can be created via the
"ALTER TABLE ... ADD COLUMN ... DEFAULT" SQL command. These columns
should point to a nested string or numbers value. ClickHouse will then
automatically fall back to this fields when the materialized column
istn't present for older data.

We also improved the bucketing, so that buckets are also created for
time ranges with less then 30 seonds.