Add support to visualize logs in ClickHouse
It is now possible to visualize logs from ClickHouse via bar or pie
charts. For that a new link to the visualization page was added to the
logs page.

To generate a pie or bar chart a user must provide a query and a time
range. Besides these two option a user can also select the operation
which should be used for the visualization, the field on which this
operation should be run, a field to group the results, a sorting order
and a limit.

NOTE: Currently the user must now if he uses a string or number field.
Maybe we can get a list of fields from ClickHouse in the future and the
let the user select the field via a select box, where we also now if the
field is a number or a string.