Remove legend from LogsChart and use useDimensions
This commit removes the legend from the LogsChart component, because on
some zoom levels the legend was displayed, which we do not want.

We are also using the useDimensions hook from the core package now,
instead the custom implementation tp determine the height and width of
the chart. For this we also added an option to specify a default for the
width and height, because the Patternfly charts package logs some errors
when the initial width / height is zero and this ways we can set it to
1.