Adjust chart handling across plugins
We moved some of the components and objects we need for the nivo chart
package to the core plugin, so that we can reuse the functions across
all our plugins. We also did the same for the colors functions which
were used by several plugins.

Besides that we improved also the styling of the LogsChart in the
ClickHouse plugin, so that it looks like all our other charts and we
renamed the visualization API endpoint to aggregations.