Add Custom Resource Definition for Users (#170)
The new Custom Resource Definition for Users can be used to define the
members of a team. For that each user can be a member of multiple teams,
which can be defined in the CRD.

With the new CRD we also added a users plugin, which can be used to view
all users or a single one and to show a list of members of a team on a
dashboard.

NOTE: At the moment the "id" field of a user isn't used, but make sure
that the id is unique accross all users in all clusters and namespace of
a kobs instance. Later the id will be used together with the
authentication feature of kobs to provide a proper authorisation
mechanism.