Add authentication and authorization mechanism
kobs now has a proper mechanism for authentication and authorization.
While the authentication of a user should still be handled via an
external application like OAuth2-Proxy, the User CR and Team CR and now
be used to authorize the authenticated user to access several plugins
and resources.

For this a User CR must be defined with the ID which is set by the
OAuth2-Proxy via the user details header. The user will then get all the
permissions from the teams he is a member of. For that the Team CR now
contains a new "permissions" property, where the plugins and resources
which can be accessed by the teams members are specified.