[istio] Add Istio Plugin
This commit adds the first version of the new Istio plugin, which can
currently show a list of applications which are part of the service
mesh. For these application we are showing the success rate, requests
per second, P50, P90 and P99 latency.

When a user selects one an application a details page for this
application is opened, where a simple topology graph for the incoming
and outgoing taffic is shown. Besides that we also show the mentioned
metrics for all versions and pods of the application.

The Istio plugin requires a configured Prometheus instance, which can be
passed to the Istio plugin in the configuration by it's name. This
Prometheus instance will then be used to get a list of all namespaces,
apllications and pods, to get the data for the React UI.

NOTE: In the next steps, we will also provide the access logs from Istio
via the plugin and some useful aggregations based on the logs. In the
third step we have to adjust the documentation and demo and we have to
add a panel for each metric.

NOTE: To use an configured plugin within another plugin, we have to
initialize all plugins first and then add them to the plugins router.
This is required, because plugins, can now also return a list of there
instances, which are then passed to other plugins (e.g. the Prometheus
plugin returns a list of instances, which is then used in the Istio
plugin to add the Prometheus instance to the Istio plugin instance).