Add sorting and show milliseconds in time column (#179)
Once a user selected another column then the default timestamp column
for sorting the logs lines, he must go to the options menu to sort the
columns for the timestamp column. Now he can directly go back to the
default by clicking the sort icon next to the time column.

The time column now uses a custom function to format the timestamp, so
that we can also display the milliseconds for the log line.