[core] Make time selection more intuitive (#181)
Sometimes the shown time from the Options component in a toolbar was not
very meaningful. Especially when a user opened a shared URL, were the
user select a last x minutes interval, we showed this text, instead of
the correct time range, also when the last x minutes referred to an very
old time.

Now we are always showing the formatted timestamp, so that this problem
can not occure anymore. It is still possible to refresh a time range, so
that the current time is used as end time with the formaly selected diff
between the start and end time.

If you are using the options component in one of your plugins, you have
to adjust the passed in values, so that your plugin works with the next
release of kobs.

NOTE: Maybe we can also remove some duplicated code for the toolbar and
the state handling via the URL in the future.