[istio] Add top and tap command and details (#182)
Next to the metrics view, the Istio plugin now also supports a top and
tap command, which used an existing Clickhouse instance. The top command
uses the Istio access logs from Clickhouse to create an aggregation
based on the path and method of an request. The tap command can be used
to view all requests. Both commands are supporting an live update
feature to refresh the view.

All three views (metrics, top and tap) are supporting details now, this
means when a user selects a metrics or a log line a drawer panel is
opened with the success rate, requests per seconde and latency metrics
for the selected time range.

NOTE: As we did it with the Prometheus plugin in #177 we are now also
exposing the Clickhouse instances from the plugin, so that we can use
the instance specified in the configuration within the Istio plugin.