[clickhouse] Add aggregations to visualize logs
It is now possible to visualize logs via aggregations. For that a new
aggregations page was added, where a user can visualize his logs via
different chart types (pie, bar, line and area).

These aggregations can now also be used within a dashboard, for that a
new "aggregation" type was added for the Clickhouse panel.

Note: This replaces the existing visualization feature.