[jaeger] Fix tooltip position in traces chart (#186)
This commit fixes the tooltip position in the traces chart shown on the
Jaeger page. It could happen that the tooltip was shown on the wrong
side, because we just compared the index of the rendered node and not
the time of the traces with the start and end time shown in the chart.
This should now be fixed.