[clickhouse] Add options to download returned logs
It is now possible to download the returned logs as ".log" or ".csv"
file. The ".csv" option is only available, when the user selected at
least one field, which should be displayed in the results table. The csv
file than contains the time column and all other selected fields as
columns.