[sql] Replace GetQueryResults function
This commit replaces the GetQueryResults function used in the SQL
plugin with the implementation we are already using in the Clickhouse
plugin, so that we have a proper handling for float values.

Background: The original function would fail, when the SQL query
returned a float value with "NaN" or "Inf" as value, because then the
json encoding would fail.