[clickhouse] Fix returned logs for custom order (#191)
There was a bug, that we returned the wrong logs, when a user selected a
custom order for the results. This was caused by our optimization of the
time range for the raw logs query by the results of the bucket query.

To fix this bug, we have adjusted the logic to adjust the time range for
the raw logs query, based on the buckets. We are using the existing
logic only for the default order "timestamp DESC". For "timestamp ASC"
order we are looping the other way around through the buckets. For each
other orders we skip the optimization based on the given limit.