[elasticsearch] Adjust time range and query (#193)
It is now possible to adjust the selected time range via the logs chart.
For that we are now using the "patternfly/react-charts" package to
render the chart.

We also added some options to adjust the query via the details view for
a single document. So that a user can directly add filter or show fields
in the results table.