[elasticsearch] Use pagination instead for results (#194)
Instead if infinite scrolling (via the load more button) we are now
using pagination to display the documents returned by our API. For that
we can remove the depracated scroll API and directly request 1000
documents from the Elasticsearch API.

Using pagination instead of infite scrolling improves the performance of
our react UI when the user wants to view a lot of documents, e.g. when
the user wants to show 500 documents, the UI becomes laggy with infinite
scrolling. This problem can be avoided with pagination.