[istio] Change upstream cluster and path handling
This commit changes the handling of the upstream cluster and path field
in the Istio access logs.

We are now using the upstream cluster in the React UI instead of the
authority. For that we also removed the authority from the grouping
condition in the SQL query for Clickhouse.

When getting the path from Clickhouse we are now ignoring the query
string.