[clickhouse] Change filters in aggregations
This commit changes the filters which can be used in aggregations to
allow more complex filters via our query syntax. With this change we can
use more then one field in a filter for comparision.

We also adjusted the intervals to not always return a datapoint every 30
seconds, instead we are returning a maximum of 100 datapoints and for
smaller intervals we are reducing the number of datapoints.