Add support for volumes and volumeMounts in Helm
When a user deploys kobs via the Helm chart it is now possible to set
additional volumes and volumeMounts in the Helm chart. This way a user
can also mount a kubeconfig and use the kubeconfig provider for kobs.

We also added the development setting to the Helm chart, so that we can
test the Helm chart without installing a Ingress Controller in a
cluster.

Fixes #198