Add trace logs for getting a dashboard logic
Because there seems to be some problems with the logic of replacing the
placeholders in a dashboard (see #200), we add more logs to the
getDashboard request. This should help us to see if we try to replace
the placeholders in a dashboard or not.

We also slightly adjusted the logic, so that a dashboard is returned
when the placeholders map is nil.