Speed up GitHub Action for the Docker image
Currently the GitHub Action for building our Docker image runs between
1 and 2 hours. To speed up this process we are now building the React UI
within the GitHub Action and not in the Docker image, so that this
only must be done once.

For that we introduced another Dockerfile which should only be used in
the GitHub action. This Dockerfile uses the same steps as the main
Dockerfile, but omits the build step for the React UI.