[resources] Add support to download/upload files
It is now possible to download/upload files from/to a container. For
that we added two new actions for Pods. The "Download File" action can
be used to download a file from a container by setting the source path
of the file from the container. The "Upload File" action can be used to
upload a file from your local machine to a container. For that the
source file must be selected and the destination path, where the file
should be created must be provided.

Note: Currently it is not possible to download a whole directory,
because we are receiving a "Truncated tar archive" for the downloaded
archive.