Add documentation about how to develop a plugin
The documentation about how to develop a plugin was missing. In the new
section "Develop a Plugin" we describe how any user can develop a plugin
for kobs. More information about plugins and how to add them to kobs can
found in the already existing section "Use the kobsio/app".