[klogs] Rename clickhouse plugin to klogs
We decided to rename the "clickhouse" plugin to "klogs".

In it's original form the "clickhouse" plugin was intended to be used
with the kobsio/klogs Fluent Bit plugin and any other ClickHouse
instance. But nowdays the clickhouse plugin was only used for logs and
raw SQL queries were moved to the "sql" plugin. Therefore we decided to
rename the plugin to "klogs".