Move plugin configuration to the plugin page
Until now the configuration for each plugin was found in the
"Configuration" section of the documentation, but it makes more sens
when the configuration for each plugin can be found on the corresponding
plugin documentation. For that we added a configuration section for each
plugin.

This means the "Configuration" section for each plugin, shows all
available options which can be set in the configuration file. The
"Options" section shows how the plugin can be used within a dashboard.