Add documentation about how to develop a plugin (#208)
* Add documentation about how to develop a plugin

The documentation about how to develop a plugin was missing. In the new
section "Develop a Plugin" we describe how any user can develop a plugin
for kobs. More information about plugins and how to add them to kobs can
found in the already existing section "Use the kobsio/app".

* Move plugin configuration to the plugin page

Until now the configuration for each plugin was found in the
"Configuration" section of the documentation, but it makes more sens
when the configuration for each plugin can be found on the corresponding
plugin documentation. For that we added a configuration section for each
plugin.

This means the "Configuration" section for each plugin, shows all
available options which can be set in the configuration file. The
"Options" section shows how the plugin can be used within a dashboard.