[techdocs] Add TechDocs Plugin
The TechDocs plugin can be used to display the documentation for your
applications and services within kobs. For that you can write your
documentation in the repository for each service and push any changes to
the documentation to an storage like S3. kobs can then render your
markdown files from a S3 bucket, so that you can access all of your
documentation within one service.

kobs also supports to embed your TechDocs into a dashboard. For that you
can choose between the "list" and "toc" type for the TechDocs plugin.
The "list" type is used to display all available TechDocs and the "toc"
type to display the table of contents for a single service.

Besides the normal markdown syntax, kobs also supports a specical code
block language named "kobs", which can be used to display a panel from
another plugin within your documentation.