[azure] Use resource groups (#217)
We are now returning all resource groups for the configured subscription
in the Azure plugin. The returned resource groups can then be passed to
all components of the Azure plugin.

That way we can now list all container instances per resource group
instead of listing all container instances in the subscription.

In the future this allows us to restrict the access for user to a
resource group instead of allowing users to access all resources in the
configured Azure subscription.