[azure] Add permissions for Azure plugin
It is now possible to restrict the permissions to resources, resource
groups and action in the Azure plugin. For that we extended the Teams
CRD with a custom permissions field, which can be used to define the
permissions for an Azure instance.

The permission schema for the Azure plugin and an example can be found
in the documentation of the Azure plugin.